Finished up conversion of the database from json to my new custom file format. The race results have now also been converted and can be read. All code for converting the files, and proof of concepts for showing how to read from them can be found in the the convert_database folder. Besides all that I also updated the RaceTime files in /src/util. The function string_to_ms used to return a long, but I found out that an unsigned int will always be enough space so I changed that. This makes it so that all time formats can be stored as unsigned int in the files instead of strings.

+++ src/api/analyzed.cpp
@@ -26,9 +26,9 @@ typedef struct
 
     char* fiscode = 0;
     char* rank = 0;
-    long time_ms = 0L;
-    long winner_time_ms = 0L;
-    long diff_ms = 0L;
+    unsigned int time_ms = 0L;
+    unsigned int winner_time_ms = 0L;
+    unsigned int diff_ms = 0L;
     float diff_percentage = 0.0f;
 
 } AnalyzedRace;
@@ -284,8 +284,8 @@ int api_getAnalyzedResults_qual(int socket, char* fiscode)
 
             // Loop though all races from the "results file"
             char* rank = 0;
-            long time = 0L;
-            long winner_time = 0L;
+            unsigned int time = 0L;
+            unsigned int winner_time = 0L;
             cJSON* races_results = cJSON_GetObjectItemCaseSensitive(JSON_race_result, "races");
             cJSON* current_race_results = NULL;
